Restructure code to conform to Python packaging norms
