+++ src/dds/text_handler.py
